[MODIFY] 1차 Merge 후 planner_date 파트 추가 수정
